backend/pyproject.toml
[tool.poetry]
name = "hangman"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "../README.md"
packages = [{ include = "src" }]

[tool.poetry.scripts]
hangman = "src.main:main"

[tool.poetry.dependencies]
python = "^3.13"
Flask = "^3.1.0"
flask-cors = "^5.0.0"
zufallsworte = "^0.0.6"
werkzeug = "^3.1.3"
pyserde = "^0.22.3"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
